Compile fixes

+++ node/src/main/scala/com/wavesplatform/database/RocksDBWriter.scala
@@ -568,7 +568,7 @@ abstract class RocksDBWriter private[database] (
         val newlyApprovedFeatures = featureVotes(height)
           .filterNot { case (featureId, _) => settings.functionalitySettings.preActivatedFeatures.contains(featureId) }
           .collect {
-            case (featureId, voteCount) if voteCount + (if (blockMeta.getHeader.featureVotes.contains(featureId)) 1 else 0) >= minVotes =>
+            case (featureId, voteCount) if voteCount + (if (blockMeta.getHeader.featureVotes.contains(featureId.toInt)) 1 else 0) >= minVotes =>
               featureId -> height
           }
